+++ kernels/include/gnuradio/kernels/common/meson.build
@@ -1,0 +1,6 @@
+headers = [
+    'registry.hpp',
+    'kernel.hpp',
+]
+
+install_headers(headers, subdir : 'gnuradio/kernels/common')
