+++ kernels/include/gnuradio/kernels/common/registry.hpp
