+++ kernels/include/gnuradio/kernels/meson.build
@@ -1,2 +1,3 @@
+subdir('common')
 subdir('cpu')
 subdir('cuda')
