+++ runtime/lib/meson.build
@@ -3,7 +3,7 @@ incdir = [
 ]
 
 
-runtime_deps = [yaml_dep, spdlog_dep, threads_dep, fmt_dep, rt_dep, pmtf_dep]
+runtime_deps = [yaml_dep, spdlog_dep, threads_dep, fmt_dep, rt_dep, pmtf_dep, newsched_kernels_cpu_dep]
 
 if cuda_dep.found() and get_option('enable_cuda')
   cu_sources = [
@@ -228,7 +228,7 @@ endif
 newsched_runtime_lib = library('newsched-runtime', 
     runtime_sources, 
     install : true, 
-    dependencies : [runtime_deps, newsched_kernels_cpu_dep],
+    dependencies : runtime_deps,
     link_language : 'cpp',
     include_directories : incdir,
     cpp_args : cpp_args
